Reviewed by Brady.
        Build fix for Xcode 2.3. Convert GraphicsContext::translate() to
        take floats. There was already an implicit conversion to float/double
        when calling CGContextTranslateCTM or cairo_translate.

        * platform/GraphicsContext.h:
        * platform/cairo/GraphicsContextCairo.cpp:
        (WebCore::GraphicsContext::translate):
        * platform/cg/GraphicsContextCG.cpp:
        (WebCore::GraphicsContext::translate):



Canonical link: https://commits.webkit.org/13876@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@16527 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/platform/GraphicsContext.h
@@ -159,7 +159,7 @@ namespace WebCore {
 
         void scale(const FloatSize&);
         void rotate(float angleInRadians);
-        void translate(int x, int y);
+        void translate(float x, float y);
         IntPoint origin();
 
         void concatCTM(const AffineTransform&);
